--- a/mujoco/mjx/_src/support.py
+++ b/mujoco/mjx/_src/support.py
@@ -37,6 +37,7 @@ def mul_m(
   d: Data,
   res: wp.array(ndim=2, dtype=wp.float32),
   vec: wp.array(ndim=2, dtype=wp.float32),
+  skip: wp.array(ndim=1, dtype=bool),
 ):
   """Multiply vector by inertia matrix."""
 
@@ -45,8 +46,12 @@ def mul_m(
     def tile_mul(adr: int, size: int, tilesize: int):
       # TODO(team): speed up kernel compile time (14s on 2023 Macbook Pro)
       @kernel(module="unique")
-      def mul(m: Model, d: Data, leveladr: int, res: array3df, vec: array3df):
+      def mul(m: Model, d: Data, leveladr: int, res: array3df, vec: array3df, skip: wp.array(ndim=1, dtype=bool)):
         worldid, nodeid = wp.tid()
+
+        if skip[worldid]:
+          return
+        
         dofid = m.qLD_tile[leveladr + nodeid]
         qM_tile = wp.tile_load(
           d.qM[worldid], shape=(tilesize, tilesize), offset=(dofid, dofid)
@@ -65,6 +70,7 @@ def mul_m(
           adr,
           res.reshape(res.shape + (1,)),
           vec.reshape(vec.shape + (1,)),
+          skip,
         ],
         # TODO(team): develop heuristic for block dim, or make configurable
         block_dim=32,
@@ -85,8 +91,13 @@ def mul_m(
       d: Data,
       res: wp.array(ndim=2, dtype=wp.float32),
       vec: wp.array(ndim=2, dtype=wp.float32),
+      skip: wp.array(ndim=1, dtype=bool),
     ):
       worldid, dofid = wp.tid()
+
+      if skip[worldid]:
+        return
+        
       res[worldid, dofid] = d.qM[worldid, 0, m.dof_Madr[dofid]] * vec[worldid, dofid]
 
     @kernel
@@ -95,8 +106,13 @@ def mul_m(
       d: Data,
       res: wp.array(ndim=2, dtype=wp.float32),
       vec: wp.array(ndim=2, dtype=wp.float32),
+      skip: wp.array(ndim=1, dtype=bool),
     ):
       worldid, elementid = wp.tid()
+
+      if skip[worldid]:
+        return
+
       i = m.qM_mulm_i[elementid]
       j = m.qM_mulm_j[elementid]
       madr_ij = m.qM_madr_ij[elementid]
@@ -106,10 +122,10 @@ def mul_m(
       wp.atomic_add(res[worldid], i, qM * vec[worldid, j])
       wp.atomic_add(res[worldid], j, qM * vec[worldid, i])
 
-    wp.launch(_mul_m_sparse_diag, dim=(d.nworld, m.nv), inputs=[m, d, res, vec])
+    wp.launch(_mul_m_sparse_diag, dim=(d.nworld, m.nv), inputs=[m, d, res, vec, skip])
 
     wp.launch(
-      _mul_m_sparse_ij, dim=(d.nworld, m.qM_madr_ij.size), inputs=[m, d, res, vec]
+      _mul_m_sparse_ij, dim=(d.nworld, m.qM_madr_ij.size), inputs=[m, d, res, vec, skip]
     )