--- a/mujoco/mjx/_src/support_test.py
+++ b/mujoco/mjx/_src/support_test.py
@@ -50,7 +50,8 @@ class SupportTest(parameterized.TestCase):
 
     res = wp.zeros((1, mjm.nv), dtype=wp.float32)
     vec = wp.from_numpy(np.expand_dims(mj_vec, axis=0), dtype=wp.float32)
-    mjx.mul_m(m, d, res, vec)
+    skip = wp.zeros((d.nworld), dtype=bool)
+    mjx.mul_m(m, d, res, vec, skip)
 
     _assert_eq(res.numpy()[0], mj_res, f"mul_m ({'sparse' if sparse else 'dense'})")