this pr aims to improve the performance of the engine solver by incorporating early returns in kernels launched by `solve`.

the first benchmarking results evaluate overhead introduced by the early returns by considering one solver iteration:

**solver=cg, iterations=1, ls_iterations=20**
```
mjwarp-testspeed --function=solve --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=cg --iterations=1 --ls_iterations=20
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 4.36 s
 Total simulation time: 3.98 s
 Total steps per second: 2,058,020
 Total realtime factor: 10,290.10 x
 Total time per step: 485.90 ns
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.10 s
 Total simulation time: 3.97 s
 Total steps per second: 2,065,483
 Total realtime factor: 10,327.42 x
 Total time per step: 484.15 ns
 ```
 

 **solver=newton, iterations=1, ls_iterations=20**
 ```
 mjwarp-testspeed --function=solve --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=newton --iterations=1 --ls_iterations=20
 ```
 
 this pr:
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 2.78 s
 Total simulation time: 8.01 s
 Total steps per second: 1,022,150
 Total realtime factor: 5,110.75 x
 Total time per step: 978.33 ns
 ```
 
 main:
 ```
 mjwarp-testspeed --function=solve --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=newton --iterations=1 --ls_iterations=20
 ```
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 0.10 s
 Total simulation time: 7.99 s
 Total steps per second: 1,024,950
 Total realtime factor: 5,124.75 x
 Total time per step: 975.66 ns
 ```
 
 **for both cg and newton solvers a small amount of overhead is introduced by the early returns**
 
 the second set of benchmarks evaluate the effect of early returns on the engine solver performance by considering multiple solver iterations:
 
 **solver=cg, iterations=10, ls_iterations=20**
 ```
 mjwarp-testspeed --function=solve --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=cg --iterations=10 --ls_iterations=20
 ```
 
 this pr:
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 4.95 s
 Total simulation time: 8.68 s
 Total steps per second: 943,349
 Total realtime factor: 4,716.74 x
 Total time per step: 1060.05 ns
 ```
 
 main:
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 0.48 s
 Total simulation time: 24.11 s
 Total steps per second: 339,711
 Total realtime factor: 1,698.56 x
 Total time per step: 2943.68 ns
 ```
 
 **solver=newton, iterations=10, ls_iterations=20**
 ```
 mjwarp-testspeed --function=solve --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=newton --iterations=10 --ls_iterations=20
 ```
 
 this pr:
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 3.29 s
 Total simulation time: 23.92 s
 Total steps per second: 342,519
 Total realtime factor: 1,712.59 x
 Total time per step: 2919.55 ns
 ```
 
 main:
 ```
 Summary for 8192 parallel rollouts

 Total JIT time: 0.47 s
 Total simulation time: 46.95 s
 Total steps per second: 174,468
 Total realtime factor: 872.34 x
 Total time per step: 5731.70 ns
 ```
 
 **for both solvers and multiple solver iterations, the early returns improve performance compared to the existing implementation**
 
 **in summary, the benchmarks demonstrate that early returns introduce a small amount of overhead for one solver iteration and can improve performance for multiple solver iterations.**