@erikfrey 2025-03-10
Let's discuss this in person along with your parallel linesearch PR - I think they're actually related and I want to talk over how we think about them and whether we would automatically switch between iterative and parallel, and how.